Created a config.sample file to configure the necessary variables for the different scripts
Most of the variables are shared by the different scripts. So the should be configured in one central file. The config.sample file is to be renamed and edited before use. It is sourced (with . config) in the different script files.

Variables not needed anymore were removed from the script files.