Run refresh_yum_repo.sh only for epel repos
Running refresh_yum_repos.sh on all REPOS to rebuild caused issues with mirrored
RHEL repos. Errata information was lost because updateinfo.xml.gz was removed
during the run of refresh_yum_repo.sh.

I changed the script that refesh_yum_repo.sh runs only on epel and epel-modular,
where it is necessary but not on the RHEL Repos where it causes trouble.