Update files for use with EPEL
While the upstream RHEL repos keep all package versions for a major release around
EPEL contains only the most current package version. When synching EPEL without
deleting older packages you have the RPMs on your mirror but are not able to use them.
You have to rebuild the repo metadata using refresh_yum_repo.sh for example.

But you cannot use refresh_yum_repo.sh on the mirrored RHEL repos because it removes
the updateinfo.xml.gz and the errata information are lost.

This commit takes care that refresh_yum_repo.sh only runs on EPEL and only when
epel or epel-modular are specified in REPOID in the CONFIG file.